Update patch set 1
Patch Set 1:

If we maintain this file as is originaly from google, makes easyer to upgrade on each google change since is only copy their changes.

This commit is a big change of the file and will ask a maintenance at each upgrade, with no benefit for the resultant ROM.

Nexus devices, since are constanly maintained from google, seams to be better to maintain their extructure unless we apply a real benefit for the enduser.

Patch-set: 1